refactor(runtime): split orchestrator/tests — extract bus + worker integration
Extract the 12 bus and worker integration tests into a new
`tests/bus_worker.rs`, combining the "Bus integration tests" and
"Worker integration tests" sections:

Bus (5):
- `parse_interface_known_values`
- `parse_interface_unknown_falls_back_to_cli`
- `run_worker_processes_turn_request`
- `submit_turn_publishes_and_waits_for_result`
- `worker_propagates_submit_correlation_fields_to_turn_result`

Worker (7):
- `submit_turn_with_worker_returns_result`
- `submit_turn_without_worker_times_out`
- `with_submit_timeout_respected`
- `failed_turn_propagates_error`
- `timeout_cancels_in_flight_worker_turn`
- `failed_turn_with_tools_llm_error_propagates`
- `failed_turn_with_tools_max_iterations_returns_error`

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); `orchestrator::tests::bus_worker::*` subtotal is 12.

Fourth slice in the stack.
