+++ crates/runtime/src/orchestrator/tests/mod.rs
@@ -1,13 +1,12 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::time::Duration;
 
 use assistant_core::tool::{ToolHandler, ToolOutput};
 use assistant_core::types::conversation::{ExecutionContext, Interface, TurnIdentity};
 use assistant_core::{
-    ChatHistoryMessage, ChatRole, ContentBlock, LlmProvider, MessageBus, OrgId, PublishRequest,
-    SpaceId, ToolCallItem, UserId, bus_messages, topic, types::agent::AssistantConfig,
+    ChatHistoryMessage, ChatRole, ContentBlock, LlmProvider, MessageBus, OrgId, SpaceId,
+    ToolCallItem, UserId, types::agent::AssistantConfig,
 };
 use assistant_llm_provider::ollama::client::{LlmClient, LlmClientConfig};
 use assistant_storage::{StorageLayer, registry::SkillRegistry};
@@ -139,6 +138,7 @@ impl ToolHandler for MockExtTool {
     }
 }
 
+mod bus_worker;
 mod end_turn;
 mod streaming;
 mod subagent;
@@ -1267,197 +1267,6 @@ fn sanitize_history_tool_result_after_matched_calls_dropped() {
     ));
 }
 
-// ── Bus integration tests ────────────────────────────────────────────────
-
-#[test]
-fn parse_interface_known_values() {
-    use super::parse_interface;
-    assert_eq!(parse_interface("Cli"), Interface::Cli);
-    assert_eq!(parse_interface("cli"), Interface::Cli);
-    assert_eq!(parse_interface("Slack"), Interface::Slack);
-    assert_eq!(parse_interface("MATTERMOST"), Interface::Mattermost);
-    assert_eq!(parse_interface("Signal"), Interface::Signal);
-    assert_eq!(parse_interface("mcp"), Interface::Mcp);
-}
-
-#[test]
-fn parse_interface_unknown_falls_back_to_cli() {
-    use super::parse_interface;
-    assert_eq!(parse_interface("unknown"), Interface::Cli);
-    assert_eq!(parse_interface(""), Interface::Cli);
-}
-
-#[tokio::test]
-async fn run_worker_processes_turn_request() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "bus response").await;
-
-    let (orch, _storage) = build(&server.uri()).await;
-
-    // Spawn the worker in the background.
-    let orch_worker = orch.clone();
-    let worker = tokio::spawn(async move {
-        orch_worker.run_worker("test-worker").await;
-    });
-
-    // Publish a TurnRequest to the bus.
-    let conv_id = Uuid::new_v4();
-    let turn_req = bus_messages::TurnRequest {
-        prompt: "hello from bus".to_string(),
-        conversation_id: conv_id,
-        extension_tools: vec![],
-        timestamp: None,
-        traceparent: None,
-        attachment_ids: vec![],
-        user_id: None,
-        org_id: None,
-        space_id: None,
-    };
-    orch.bus()
-        .publish(
-            PublishRequest::new(
-                topic::TURN_REQUEST,
-                serde_json::to_value(&turn_req).unwrap(),
-            )
-            .with_agent_id("default")
-            .with_conversation_id(conv_id)
-            .with_interface("Cli"),
-        )
-        .await
-        .unwrap();
-
-    // Poll for the worker to process and publish the result instead of
-    // a fixed sleep, which can be flaky under CI load.
-    let deadline = tokio::time::Instant::now() + Duration::from_secs(5);
-    let results = loop {
-        let r = orch.bus().list(topic::TURN_RESULT, None, 10).await.unwrap();
-        if !r.is_empty() {
-            break r;
-        }
-        assert!(
-            tokio::time::Instant::now() < deadline,
-            "timed out waiting for TurnResult"
-        );
-        tokio::time::sleep(Duration::from_millis(50)).await;
-    };
-
-    assert_eq!(results.len(), 1, "expected one TurnResult on the bus");
-    let result: bus_messages::TurnResult =
-        serde_json::from_value(results[0].payload.clone()).unwrap();
-    assert_eq!(result.conversation_id, conv_id);
-    assert_eq!(result.content, "bus response");
-
-    // The original request should be acked (done).
-    let pending = orch
-        .bus()
-        .list(
-            topic::TURN_REQUEST,
-            Some(assistant_core::MessageStatus::Pending),
-            10,
-        )
-        .await
-        .unwrap();
-    assert!(pending.is_empty(), "turn request should be acked");
-
-    worker.abort();
-}
-
-#[tokio::test]
-async fn submit_turn_publishes_and_waits_for_result() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "submitted answer").await;
-
-    let (orch, _storage) = build(&server.uri()).await;
-
-    // Spawn the worker so it can process the submitted turn.
-    let orch_worker = orch.clone();
-    tokio::spawn(async move {
-        orch_worker.run_worker("test-worker").await;
-    });
-
-    let conv_id = Uuid::new_v4();
-    let result = orch
-        .submit_turn("hello via submit", conv_id, Interface::Cli, None)
-        .await
-        .unwrap();
-    assert_eq!(result.answer, "submitted answer");
-}
-
-#[tokio::test]
-async fn worker_propagates_submit_correlation_fields_to_turn_result() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "correlated response").await;
-
-    let (orch, _storage) = build(&server.uri()).await;
-
-    let orch_worker = orch.clone();
-    let worker = tokio::spawn(async move {
-        orch_worker.run_worker("test-worker").await;
-    });
-
-    let conv_id = Uuid::new_v4();
-    let request_id = Uuid::new_v4();
-    let turn_req = bus_messages::TurnRequest {
-        prompt: "hello correlation".to_string(),
-        conversation_id: conv_id,
-        extension_tools: vec![],
-        timestamp: None,
-        traceparent: None,
-        attachment_ids: vec![],
-        user_id: None,
-        org_id: None,
-        space_id: None,
-    };
-
-    let request_msg_id = orch
-        .bus()
-        .publish(
-            PublishRequest::new(
-                topic::TURN_REQUEST,
-                serde_json::to_value(&turn_req).unwrap(),
-            )
-            .with_agent_id("default")
-            .with_conversation_id(conv_id)
-            .with_interface("Cli")
-            .with_correlation_id(request_id)
-            .with_batch_id(request_id),
-        )
-        .await
-        .unwrap();
-
-    let deadline = tokio::time::Instant::now() + Duration::from_secs(5);
-    let results = loop {
-        let r = orch.bus().list(topic::TURN_RESULT, None, 10).await.unwrap();
-        if !r.is_empty() {
-            break r;
-        }
-        assert!(
-            tokio::time::Instant::now() < deadline,
-            "timed out waiting for correlated TurnResult"
-        );
-        tokio::time::sleep(Duration::from_millis(50)).await;
-    };
-
-    let result_msg = &results[0];
-    assert_eq!(
-        result_msg.batch_id,
-        Some(request_id),
-        "TURN_RESULT should preserve the request batch_id"
-    );
-    assert_eq!(
-        result_msg.correlation_id,
-        Some(request_id),
-        "TURN_RESULT should preserve the request correlation_id"
-    );
-    assert_eq!(
-        result_msg.causation_id,
-        Some(request_msg_id),
-        "TURN_RESULT should set causation_id to the TURN_REQUEST message id"
-    );
-
-    worker.abort();
-}
-
 // ── value_to_params_map tests ─────────────────────────────────────────────────
 
 #[test]
@@ -1540,318 +1349,6 @@ async fn run_turn_max_iterations_returns_error() {
                 "error should mention max iterations: {err_msg}"
             );
         }
-    }
-}
-
-// ── Worker integration tests ────────────────────────────────────────────────
-
-/// Test that submit_turn works correctly when the worker is running.
-/// This reproduces the bug where Slack/Mattermost-only modes didn't spawn
-/// the worker, causing submit_turn to timeout waiting for a result.
-#[tokio::test]
-async fn submit_turn_with_worker_returns_result() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "hello from worker").await;
-
-    let (orch, _) = build(&server.uri()).await;
-    let conv_id = Uuid::new_v4();
-
-    // Spawn the worker in the background (as the CLI would do).
-    let worker_orch = orch.clone();
-    let _worker = tokio::spawn(async move {
-        worker_orch.run_worker("test-worker").await;
-    });
-
-    // Give the worker a moment to start.
-    tokio::time::sleep(Duration::from_millis(50)).await;
-
-    let result = orch
-        .submit_turn("test message", conv_id, Interface::Slack, None)
-        .await;
-
-    assert!(
-        result.is_ok(),
-        "submit_turn should succeed when worker is running"
-    );
-    assert_eq!(result.unwrap().answer, "hello from worker");
-}
-
-/// Test that submit_turn times out when no worker is running.
-/// This verifies the bug condition that was fixed.
-#[tokio::test]
-async fn submit_turn_without_worker_times_out() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "this should not be received").await;
-
-    let (orch, _) = build(&server.uri()).await;
-    let conv_id = Uuid::new_v4();
-
-    // No worker spawned — simulates the bug condition.
-    let result = tokio::time::timeout(
-        Duration::from_secs(2),
-        orch.submit_turn("test message", conv_id, Interface::Slack, None),
-    )
-    .await;
-
-    assert!(
-        result.is_err() || result.unwrap().is_err(),
-        "submit_turn should fail/timeout when worker is not running"
-    );
-}
-
-/// Test that with_submit_timeout sets a custom deadline that is respected.
-///
-/// We configure a 1-second timeout and verify that submit_turn (with no worker
-/// running) completes within a short wall-clock window — i.e. it honours the
-/// configured value rather than the 3-hour default.
-#[tokio::test]
-async fn with_submit_timeout_respected() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "not used").await;
-
-    // Build without wrapping in Arc so we can apply with_submit_timeout first.
-    let mut config = AssistantConfig::default();
-    config.memory.enabled = false;
-    let storage = Arc::new(StorageLayer::new_in_memory().await.unwrap());
-    let registry = Arc::new(SkillRegistry::new(storage.pool.clone()).await.unwrap());
-    let llm: Arc<dyn LlmProvider> = Arc::new(
-        LlmClient::new(LlmClientConfig {
-            model: "test".to_string(),
-            base_url: server.uri(),
-            timeout_secs: 10,
-            retry_config: assistant_llm_provider::retry::RetryConfig::disabled(),
-        })
-        .unwrap(),
-    );
-    let executor = Arc::new(ToolExecutor::new(
-        storage.clone(),
-        llm.clone(),
-        registry.clone(),
-        Arc::new(config.clone()),
-    ));
-    let bus: Arc<dyn MessageBus> = Arc::new(storage.message_bus());
-    let orch = Arc::new(
-        Orchestrator::new(llm, storage, executor.clone(), registry, bus, &config)
-            .with_submit_timeout(1), // 1-second deadline
-    );
-    executor.set_subagent_runner(orch.clone());
-
-    let conv_id = Uuid::new_v4();
-    let start = std::time::Instant::now();
-    // Wrap in a generous outer timeout to prevent the test hanging if the
-    // inner deadline is accidentally the 3-hour default.
-    let result = tokio::time::timeout(
-        Duration::from_secs(10),
-        orch.submit_turn("test message", conv_id, Interface::Slack, None),
-    )
-    .await;
-    let elapsed = start.elapsed();
-
-    // The inner submit_turn must have timed out (no worker running).
-    assert!(
-        result.is_ok() && result.unwrap().is_err(),
-        "submit_turn should return Err when no worker is running"
-    );
-    // And it should have done so close to the configured 1-second deadline.
-    assert!(
-        elapsed < Duration::from_secs(5),
-        "submit_turn should respect the 1-second timeout, not the 3-hour default (elapsed: {elapsed:?})"
-    );
-}
-
-#[tokio::test]
-async fn failed_turn_propagates_error() {
-    // Verifies that when the LLM returns an error, run_turn propagates Err.
-    // The error path is also where we set OtelStatus::Error on the turn span.
-    let server = MockServer::start().await;
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(ResponseTemplate::new(500).set_body_string("server error"))
-        .mount(&server)
-        .await;
-
-    let (orch, _storage) = build(&server.uri()).await;
-    let conv_id = Uuid::new_v4();
-
-    let result = orch
-        .run_turn(
-            "test",
-            conv_id,
-            Interface::Cli,
-            None,
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await;
-    assert!(
-        result.is_err(),
-        "expected run_turn to return Err on LLM failure"
-    );
-}
-
-/// Verify that when submit_turn times out it cancels the in-flight worker
-/// turn via the CancellationToken registered in turn_cancellations.
-///
-/// The mock LLM is configured to delay its response by 5 s.  submit_turn
-/// is given a 1-second deadline.  We assert:
-///  1. submit_turn returns an error within ~2 s (the deadline fires).
-///  2. The worker does NOT continue processing: if it did it would produce a
-///     turn.result message that would linger on the bus, which we check for.
-#[tokio::test]
-async fn timeout_cancels_in_flight_worker_turn() {
-    let server = MockServer::start().await;
-    // Delay the LLM response so the turn is still running when submit_turn times out.
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200)
-                .set_delay(Duration::from_secs(5))
-                .set_body_json(ollama_answer("delayed answer")),
-        )
-        .mount(&server)
-        .await;
-
-    let mut config = AssistantConfig::default();
-    config.memory.enabled = false;
-    let storage = Arc::new(StorageLayer::new_in_memory().await.unwrap());
-    let registry = Arc::new(SkillRegistry::new(storage.pool.clone()).await.unwrap());
-    let llm: Arc<dyn LlmProvider> = Arc::new(
-        LlmClient::new(LlmClientConfig {
-            model: "test".to_string(),
-            base_url: server.uri(),
-            timeout_secs: 30,
-            retry_config: assistant_llm_provider::retry::RetryConfig::disabled(),
-        })
-        .unwrap(),
-    );
-    let executor = Arc::new(ToolExecutor::new(
-        storage.clone(),
-        llm.clone(),
-        registry.clone(),
-        Arc::new(config.clone()),
-    ));
-    let bus: Arc<dyn MessageBus> = Arc::new(storage.message_bus());
-    let orch = Arc::new(
-        Orchestrator::new(
-            llm,
-            storage.clone(),
-            executor.clone(),
-            registry,
-            bus,
-            &config,
-        )
-        .with_submit_timeout(1), // 1-second deadline
-    );
-    executor.set_subagent_runner(orch.clone());
-
-    // Spawn the worker so it picks up and processes the turn.
-    let worker_orch = orch.clone();
-    let _worker = tokio::spawn(async move {
-        worker_orch.run_worker("test-worker").await;
-    });
-    tokio::time::sleep(Duration::from_millis(50)).await;
-
-    let conv_id = Uuid::new_v4();
-    let start = std::time::Instant::now();
-    let result = orch
-        .submit_turn("slow task", conv_id, Interface::Slack, None)
-        .await;
-    let elapsed = start.elapsed();
-
-    // submit_turn must have timed out with an error.
-    assert!(result.is_err(), "submit_turn should error on timeout");
-    assert!(
-        elapsed < Duration::from_secs(4),
-        "submit_turn should have stopped within ~1 s, not waited for the 5-s LLM response (elapsed: {elapsed:?})"
-    );
-
-    // Give the worker a moment — if cancellation worked there should be no
-    // lingering turn.result message on the bus.
-    tokio::time::sleep(Duration::from_millis(200)).await;
-    let leftover = storage
-        .message_bus()
-        .claim_filtered(
-            assistant_core::topic::TURN_RESULT,
-            "check",
-            &assistant_core::ClaimFilter::new(),
-        )
-        .await
-        .unwrap();
-    assert!(
-        leftover.is_none(),
-        "worker should not have published a turn.result after cancellation"
-    );
-}
-
-/// `run_turn_with_tools` uses `run_turn_with_tools_impl` which has its own
-/// LLM call site.  Verify that an HTTP-500 from the LLM propagates as `Err`
-/// through the extension-tool path (separate from `run_turn_core`).
-#[tokio::test]
-async fn failed_turn_with_tools_llm_error_propagates() {
-    let server = MockServer::start().await;
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(ResponseTemplate::new(500).set_body_string("server error"))
-        .mount(&server)
-        .await;
-
-    let (orch, _) = build(&server.uri()).await;
-
-    let result = orch
-        .run_turn_with_tools(
-            "test",
-            Uuid::new_v4(),
-            Interface::Slack,
-            vec![],
-            None,
-            vec![],
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await;
-    assert!(
-        result.is_err(),
-        "run_turn_with_tools must return Err on LLM HTTP-500"
-    );
-}
-
-/// `run_turn_with_tools` must return `Err` when the LLM loops forever
-/// (always returns tool calls, never a final answer).
-#[tokio::test]
-async fn failed_turn_with_tools_max_iterations_returns_error() {
-    let server = MockServer::start().await;
-    // LLM never produces a final answer — always returns a tool call.
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(ollama_tool_calls(&["unknown-tool"])),
-        )
-        .mount(&server)
-        .await;
-
-    let mut config = AssistantConfig::default();
-    config.memory.enabled = false;
-    config.llm.max_iterations = 2;
-    let (orch, _) = build_with_config(&server.uri(), config).await;
-
-    let result = orch
-        .run_turn_with_tools(
-            "loop forever",
-            Uuid::new_v4(),
-            Interface::Cli,
-            vec![],
-            None,
-            vec![],
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await;
-    match result {
-        Ok(_) => panic!("should fail when max iterations reached"),
-        Err(e) => assert!(
-            e.to_string().contains("Max iterations"),
-            "error should mention max iterations: {e}"
-        ),
     }
 }
